settings: add support for sections
We can now put key-value pairs into sections. This allows for easier,
faster and lookups when we have a load of key-value pairs that we want
to read all at once.

The HashTable now has a union as the key type - either a String key, or
a String section + String key.