add default CC to parameter mappings
Add a new boolean setting for enabling default to parameter mappings
such as CC 7 to Master Volume. When turned on, new instances of Floe
will start with these parameters 'MIDI CC learned'.

We used to have a similar functionality where if we detected that the
settings file was newly created, we would write these mappings into it.

This implementation is better though because it can be disabled, and
it can be reinstated - whereas before it relied on the settings file
being non-existent.