rename Settings to Preferences
Preferences better covers what this is - the option to prefer to use
something other than the default. Preferences seems to also be more
common term for this certainly on macOS, but on DAWs on Windows too.

Familiarity and precision of wording is important for a smooth UX.

Additional, we can use the nice, commonly understood abbreviation:
prefs.