preferences: fix incorrect reliance on file last modified change
Linux seems to lazily set the last modified time of a file, so when
querying the value right after closing a file it hasn't updated. This
messes up our polling tests. To fix this, we manually set the time
in a separate call, this prompts the OS to make it happen.