fix stacktrace start point
On Windows we need another frame. We had a hard time with this Result()
constructor before where we couldn't correctly get the program counter
of the caller. Hence why we are using a plain number of frames to skip.
Ideally we should get the bottom of this, but for time's sake let's just
do a quick fix. Having more frames doesn't do much harm, it's just a
little harder to read where the actual error came from.