ensure relative paths in stacktraces
On Linux, despite using -fmacro-prefix-map -fdebug-prefix-map and
-ffile-prefix-map, we still sometimes get absolute paths - somehow on
some files the absolute prefix is added back on after the prefix map is
applied. I know this because if we use prefix-map to create a nonsense
name, we can see the nonsense name in the output - in the middle of the
path.