big update on documentation
Improve writing and documentation everywhere: fill in gaps, be clearer,
add more images.

For the readme we have done away with the list of modules - I think for
now it's apparent enough from the code structure itself. It means we no
longer have to keep the readme in sync with the actual code.