fix tracy initialisation
We need to track init-count for Tracy. We can't put a global variable
in the header without causing the variable to be duplicated in different
translation units, so we create a cpp file for global. This is a good
change anyways. I think the 'header-only' approach for the entire
codebase is not a successful experiment due to global variable problems,
and long compile times when editing implementation details in a header.

Anyways, this patch also fixes the initialisation order so that Tracy
works again. We needed to init Tracy earlier in the chain.