support timeouts in MutexThin and use it protect logger
The logger can deadlock if there's recursion. We want to take the
approach that all 'boundless' operations should have a limit. This
includes mutex locks. We know that the logger write operation is very
quick and deterministic - so we say that 2 seconds means there's a
deadlock and we should abort.