windows: fix file dialogs
After many hours of debugging...

IFileDialog::Show() will block if the COM mode is not single
threaded apartment. There's no error message. Just infinite stall.

As a plugin, we don't know how the host might have initialised COM.
If it's not in the right mode, we have to spin up another thread
where we can guarantee correct COM initialisation. We also need
to pump messages ourselves because the dialog will indefinitely stall.

It doesn't seem ideal to be pumping messages ourselves - we don't know
what might be in the message queue and how the host might respond to it.
However, I see no alternative.