make Windows file picker async and not pump messages
We don't want to manually pump messages like we were in the async case
because we don't know the state of the host application - can it handle
this kind of recursive message pumping (receiving messages while
already in a WndProc)? The answer is probably yes, but it's not a
definite.

So instead, we now always start a new thread for the IFileDialog. It's
guaranteed to have the correct COM init. And we take a fully
non-blocking approach, where we do not pump messages ourselves but
allow whatever message pump is already running to handle the necessary
messages that result from a IFileDialog - some messages are required to
be handled on the IFileDialog's parent before IFileDialog will show the
dialog.

Additionally, the GUI now stays responsive while the dialog is open,
though I believe not interactable due to window ordering (this is
standard Windows UX, the dialog always shows on top of its parent).

To allow for this extra booking of the dialog thread, we now store a
larger opaque 'handle' on the GuiPlatfrom rather than just a void *.
This way we don't have do an additional allocation and
detroy()/desctructor.