tidy up objective-c naming a little
The Objective-C runtime uses a global namespace for all symbols it
loads. So, for an audio plugin we need to careful that other plugins
don't have colliding names - different versions of a class that has the
same name. We append a unique version to objective-c symbols to fix
this.

We were mostly doing this already but this tidies it up.