ci: fix not reporting error code
Wrong approach - we want the exit code of the command rather than the
the stdout.

All this kind of bash faff really makes it convincing to set up Zig to
do all this scripting. Zig will be much more verbose but I doubt we'll
have all this minor bash errors. In this case, this sort of thing would
be found using a type system at compile time. Additionally, we have had
lots of faff with bash on Windows - trying to reuse scripts across Unix
and Windows.