windows: add keyboard hook to ensure we get text input in DAW
DAWs often eat keyboard messages before they get to the plugin so that
their keyboard shortcuts still work. This is good, except when we have a
text input selected - the plugin needs the characters. We install a
keyboard hook to consume the messages. This is the same as JUCE does it.