mac: try fix FSEvent stream threading issue
It seems like its possible our event callback is running at the same
time as we invalidate the stream memory causing corruption.

https://github.com/Floe-Project/Floe/actions/runs/13907734074/job/38914608051

It's tricky to reproduce. But this new method seems correct. We now
dispatch the ending of the stream to the same queue that has the events
to ensure that no events are currently running.