remove requirement for packages to end .floe.zip
I don't think it adds much utility. We tend to make sure that the word
'package' is in the filename.

Additionally, it becomes a problem when redownloading files on Windows.
If the file already exists it's renamed file(1).zip. The (1) is inserted
after the .floe and before the .zip therefore making the file an invalid
package.